feat(agentboard): merge detail panel into session card
Fold the agent detail rows directly into each SessionCard so the sidebar
shows everything in one list — no more separate detail panel with its own
focus model, height config, or resize plumbing.

Also fixes a bug where closed tmux panes would leave stale agent entries
in the sidebar: mergeAgentsWithPanePresence now drops non-terminal agents
whose pane is no longer in the scan (terminals keep showing until pruned).

- Delete DetailPanel.tsx and detail-panel-height.ts
- Extract cache-bar rendering into components/cache-bar.ts
- Inline agent rows (with cache countdown bar) into SessionCard
- Support jump-to-top/bottom reordering via Alt+Shift+Up/Down
- Remove dead SPARK_BLOCKS / TONE_ICONS exports and detailPanelHeights config
