+++ packages/agentboard/README.md
@@ -83,32 +83,27 @@ WebSocket server on `127.0.0.1:4201`. Auto-started by the TUI or tmux scripts.
 Solid.js app rendered via OpenTUI. Connects to server over WebSocket.
 
 - Session cards with accent bars, status icons, branch info
-- Resizable detail panel with agent list + metadata
-- Mouse support (click, drag to resize)
+- Inline agent rows per card with cache-countdown bar for Claude Code panes
+- Mouse support (click to focus, dismiss)
 - Help overlay (`?`)
 
 #### TUI Components
 
-**`SessionCard`** (`components/SessionCard.tsx`) — session list item
+**`SessionCard`** (`components/SessionCard.tsx`) — session list item with inline agent rows
 
 - Row 1: session name (truncated to 18 chars) + status icon (braille spinner when running, `●` for unseen terminal states)
-- Row 2: git branch + listening port hint (`⌁4201`, or `⌁4201+2` for multiple)
-- Row 3: metadata summary (status text + progress like `3/5` or `42%`)
+- Row 2: git branch
+- Row 3: git diff stats
+- Row 4: metadata summary (status text + progress like `3/5` or `42%`)
+- Agent rows (one per pane): status icon + name + status text + dismiss `✕`, thread name, and for Claude Code agents a `model · cache ▰▰▱…` drain-down bar
 - Left accent bar colored by state: green (current), yellow (running), red (error), peach (interrupted), lavender (focused), teal (unseen done)
 
-**`DetailPanel`** (`components/DetailPanel.tsx`) — expanded view for focused session
-
-- Drag-resizable separator (height persisted per session in config)
-- Truncated working directory
-- Agent list via `AgentListItem` sub-component (see below)
-- Metadata section: status line with tone icon + progress, last 8 log entries
-
-**`AgentListItem`** (inside `DetailPanel.tsx`) — single agent instance row
+**`AgentRow`** (inside `SessionCard.tsx`) — single agent instance row
 
 - Status icon: braille spinner (running), `◉` (waiting), `✓` (done), `✗` (error), `⚠` (interrupted)
 - Agent name, thread name, status text
 - Dismiss `✕` button (hover turns red), click row to focus the agent's tmux pane
-- Flash animation on click
+- Flash animation on click, surface0 highlight when keyboard-focused
 
 **`HelpOverlay`** (inline in `index.tsx`) — modal overlay
 
@@ -117,9 +112,9 @@ Solid.js app rendered via OpenTUI. Connects to server over WebSocket.
 
 #### TUI Utilities
 
-- `constants.ts` — shared icons (`SPINNERS`, `UNSEEN_ICON`, `TONE_ICONS`), spark blocks for sparkline charts, theme list, tone-to-color mapping
+- `constants.ts` — shared icons (`SPINNERS`, `UNSEEN_ICON`), theme list, tone-to-color mapping
 - `mux-context.ts` — tmux detection, pane refocus after startup, client TTY and session name resolution
-- `detail-panel-height.ts` — per-session detail panel height persistence (min 4 rows, default 10)
+- `components/cache-bar.ts` — cache-countdown bar helpers (`cacheBar`, `cacheBarColor`, `shortModel`)
 
 ## Configuration
